Persistent process stuff.
Renamed metaclass and definer persistent-process to persistent-process-class.
Renamed standard-persistent-process to persistent-process, the base class for all persistent processes.
Factored out parts that do not strictly related to persistent processes and can be reused in standard processes.
This simplifies the common parts of the user interface for continuation based processes independent of being persistent or standard.